Support firefox 46+
Firefox recently added WebExtensions addon API which allow porting Chrome
extensions easily to Firefox.
This extension works on Firefox 46 with a minor addition to the
manifest.json file so Firefox is able to load it.

WebExtensions is implemented since Firefox 45 but the required methods
used here are available since Firefox 46.